requirements.txt
Scrapy==2.5.1
python-dotenv==0.19.2
beautifulsoup4==4.10.0
scrapy-selenium==0.0.7
selenium==4.1.0
